compositor|internal: Split source validation from source enumeration
`assertSourceIsReadable` takes its own module. It checks a declaration; `enumerateSource` walks a tree. They were production-consumed peers in a file named for one of them, and `resolveCatalog` imported both.

`compareStrings` and `isMissingFile` move to `portable/`. The first existed as three verbatim copies -- in `enumerateSource`, in `resolveCatalog`, and in the samples round-trip test -- each a comparator the determinism guarantee depends on reproducing identically. The second is what the enumerate/validate split would otherwise have duplicated.

Both arrive with tests. `isMissingFile` had a branch no test reached, which a two-hundred-line module's aggregate coverage had been averaging away; `compareStrings` now pins the property that made it worth writing, that its order comes from code units rather than from whatever locale data the runtime ships.

`PlanViolation` aliases the shared `Violation` rather than re-declaring the same two fields, which is what `CatalogViolation` already did.

`buildSource` moves to `resolution/test-utils/`, now that the enumeration and readability suites both build a fixture tree.
